requirements.txt
djangorestframework
asgiref==3.2.0
asn1crypto==0.24.0
astroid==2.2.5
attrs==19.1.0
autobahn==19.8.1
Automat==0.7.0
automium==0.2.6
automium-web==0.1.1
autopep8==1.4.4
backcall==0.1.0
bleach==3.1.0
blueprint==3.4.2
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
click==6.7
cryptography==2.8
Django==3.0
Flask==1.0
Flask-Cors==3.0.8
Flask-SSLify==0.1.5
gunicorn==19.9.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.3.0
incremental==17.5.0
ipykernel==5.1.3
ipython==7.10.2
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==0.24
jedi==0.15.1
Jinja2==2.9.6
jsmin==2.2.2
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.1
lazy-object-proxy==1.4.1
legacy==0.1.6
MarkupSafe==1.0
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.0.2
nbconvert==5.6.1
nbformat==4.4.0
notebook==6.0.2
oauthlib==3.0.1
pandocfilters==1.4.2
parso==0.5.2
pep8==1.7.1
pickleshare==0.7.5
Pillow==6.1.0
prometheus-client==0.7.1
prompt-toolkit==3.0.2
psycopg2==2.8.3
pyasn1==0.3.7
pycodestyle==2.5.0
pycparser==2.19
Pygments==2.5.2
PyHamcrest==1.9.0
pylint==2.3.1
pyOpenSSL==19.0.0
pyrsistent==0.15.6
python-dateutil==2.8.1
python-decouple==3.1
python3-openid==3.1.0
pytz==2019.1
pywin32==227
pywinpty==0.5.7
pyzmq==18.1.1
requests==2.22.0
requests-oauthlib==1.2.0
Send2Trash==1.5.0
six==1.12.0
sqlparse==0.3.0
<<<<<<< HEAD
stripe==2.37.1
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
traitlets==4.3.3
txaio==18.8.1
typed-ast==1.4.0
urllib3==1.25.3
virtualenv==16.6.1
virtualenvwrapper-win==1.2.5
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.12.2
wrapt==1.11.2
zipp==0.6.0
zope.interface==4.6.0
=======
Werkzeug==0.15.3
>>>>>>> 5013f2edbe1e629269621ad9988e0f4740b7dc29
